+++ services/supervisor/examples/configuration_change.rs
@@ -21,7 +21,7 @@ use exonum_merkledb::{
     access::{Access, AccessExt, FromAccess},
     Entry, ObjectHash,
 };
-use exonum_rust_runtime::{Service, ServiceFactory};
+use exonum_rust_runtime::{spec::Spec, Service};
 use exonum_testkit::{TestKit, TestKitBuilder};
 
 use exonum_supervisor::{ConfigPropose, ConfigVote, Configure, Supervisor, SupervisorInterface};
@@ -74,19 +74,13 @@ impl Configure for ConfigChangeService {
 }
 
 fn main() {
-    let service = ConfigChangeService;
-    let artifact = service.artifact_id();
-
     // Create testkit instance with our test service and supervisor.
+    let service = Spec::new(ConfigChangeService).with_instance(SERVICE_ID, SERVICE_NAME, ());
     let mut testkit = TestKitBuilder::validator()
         .with_logger()
         .with_validators(4)
-        .with_rust_service(Supervisor)
-        .with_artifact(Supervisor.artifact_id())
-        .with_instance(Supervisor::decentralized())
-        .with_artifact(artifact.clone())
-        .with_instance(artifact.into_default_instance(SERVICE_ID, SERVICE_NAME))
-        .with_rust_service(service)
+        .with(Supervisor::decentralized())
+        .with(service)
         .build();
 
     // Firstly, lets change consensus configuration and increase `min_propose_timeout`.